Create copy for step 7
Made a copy of smaller PASS image for step 7 into the step 7 folder; not sure why the smaller PASS image is still getting expanded to full size in the Introduction page (and now looks awful), but it's working in step 7.

Also changing intro to point to full size for now, for PASS and the new product life cycle image.